chore: update cargo for todoapp (#245)
* update cargo for todoapp

* chore(pre-commit.ci): auto fixes from pre-commit hooks

* fix: add features for cot dependencies

* chore: update edition in examples package

* chore(pre-commit.ci): auto fixes from pre-commit hooks

* chore: removed unused code

---------
