+++ examples/admin/src/migrations.rs
@@ -1,6 +1,6 @@
 //! List of migrations for the current app.
 //!
-//! Generated by cot CLI 0.1.3 on 2025-03-12 13:50:59+00:00
+//! Generated by cot CLI 0.2.0 on 2025-03-28 19:43:54+00:00
 
 pub mod m_0001_initial;
 /// The list of migrations for current app.
