+++ examples/admin/templates/index.html
@@ -9,7 +9,24 @@
 </head>
 <body>
 <h1>Hello!</h1>
-<p>Go to the <a href="{{ cot::reverse!(urls, "cot_admin:login")? }}">admin panel</a>.</p>
+
+{% if todo_items.is_empty() %}
+<p>There are no TODO items.</p>
+{% else %}
+<ul id="todo-list">
+    {% for todo in todo_items %}
+    <li>
+        {% let todo_id = todo.id %}
+        <form action="{{ cot::reverse!(urls, "remove-todo", todo_id = todo_id)? }}" method="post">
+        <span>{{ todo.title }}</span>
+        <button type="submit">Remove</button>
+        </form>
+    </li>
+    {% endfor %}
+</ul>
+{% endif %}
+
+<p>Go to the <a href="{{ cot::reverse!(urls, "cot_admin:login")? }}">admin panel</a> to manage the TODO items.</p>
 <p>The username is <strong>admin</strong> and the password is <strong>admin</strong>.</p>
 </body>
 </html>
