+++ examples/custom-task/Cargo.toml
@@ -3,7 +3,7 @@ name = "example-custom-task"
 version = "0.1.0"
 publish = false
 description = "Custom tasks - Cot example."
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 async-trait = "0.1"
