+++ examples/json/Cargo.toml
@@ -3,7 +3,7 @@ name = "example-json"
 version = "0.1.0"
 publish = false
 description = "JSON - Cot example."
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 cot = { path = "../../cot" }
