+++ examples/sessions/Cargo.toml
@@ -3,7 +3,7 @@ name = "example-sessions"
 version = "0.1.0"
 publish = false
 description = "Sessions - Cot example."
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 cot = { path = "../../cot" }
