+++ examples/sessions/src/main.rs
@@ -7,7 +7,7 @@ use cot::request::Request;
 use cot::response::{Response, ResponseExt};
 use cot::router::{Route, Router, Urls};
 use cot::session::Session;
-use cot::{reverse_redirect, App, AppBuilder, Body, BoxedHandler, Project, StatusCode};
+use cot::{App, AppBuilder, Body, BoxedHandler, Project, StatusCode, reverse_redirect};
 use rinja::Template;
 
 #[derive(Debug, Template)]
