+++ examples/todo-list/Cargo.toml
@@ -3,7 +3,7 @@ name = "example-todo-list"
 version = "0.1.0"
 publish = false
 description = "TODO List - Cot example."
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 cot = { path = "../../cot" }
