+++ examples/todo-list/src/main.rs
@@ -4,14 +4,14 @@ use cot::auth::db::DatabaseUserApp;
 use cot::cli::CliMetadata;
 use cot::config::{DatabaseConfig, ProjectConfig};
 use cot::db::migrations::SyncDynMigration;
-use cot::db::{model, query, Auto, Model};
+use cot::db::{Auto, Model, model, query};
 use cot::form::Form;
 use cot::project::{MiddlewareContext, RegisterAppsContext};
 use cot::request::extractors::{Path, RequestDb, RequestForm};
 use cot::response::{Response, ResponseExt};
 use cot::router::{Route, Router, Urls};
 use cot::static_files::StaticFilesMiddleware;
-use cot::{reverse_redirect, App, AppBuilder, Body, BoxedHandler, Project, StatusCode};
+use cot::{App, AppBuilder, Body, BoxedHandler, Project, StatusCode, reverse_redirect};
 use rinja::Template;
 
 #[derive(Debug, Clone)]
