--- examples/todoapp/.gitignore
@@ -1,15 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-debug/
-target/
-
-# These are backup files generated by rustfmt
-**/*.rs.bk
-
-# MSVC Windows builds of rustc generate these, which store debugging information
-*.pdb
-
-# Test databases
-*.db
-*.sqlite3
-*.sqlite3-journal
