--- examples/todoapp/bacon.toml
@@ -1,5 +1,0 @@
-[jobs.serve]
-command = ["cargo", "run"]
-background = false
-on_change_strategy = "kill_then_restart"
-watch = ["templates", "static"]
