--- examples/todoapp/config/prod.toml.example
@@ -1,7 +1,0 @@
-secret_key = ""  # Set this to a random hex string in production
-
-[database]
-url = "sqlite://db.sqlite3?mode=rwc"
-# Or:
-# url = "postgres://user:password@localhost:5432/database"
-# url = "mysql://user:password@localhost:3306/database"
