--- examples/todoapp/rustfmt.toml
@@ -1,7 +1,0 @@
-format_code_in_doc_comments = true
-group_imports = "StdExternalCrate"
-imports_granularity = "module"
-normalize_comments = true
-unstable_features = true
-use_field_init_shorthand = true
-wrap_comments = true
